process_emails/requirements.txt
annotated-types==0.7.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudevents==1.11.0
Deprecated==1.2.18
deprecation==2.1.0
docstring_parser==0.16
Flask==3.1.0
functions-framework==3.8.2
google-api-core==2.24.2
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.83.0
google-cloud-appengine-logging==1.6.0
google-cloud-audit-log==0.3.1
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-cloud-error-reporting==1.11.1
google-cloud-logging==3.11.4
google-cloud-resource-manager==1.14.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
grpc-google-iam-v1==0.14.1
grpcio==1.71.0
grpcio-status==1.62.3
gunicorn==23.0.0
html2text==2024.2.26
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.3
oauthlib==3.2.2
opentelemetry-api==1.30.0
packaging==24.2
proto-plus==1.26.1
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==76.0.0
shapely==2.0.7
six==1.17.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
